Move Pascal AST into shared ast module
