+++ src/Pascal/opt.h
@@ -1,7 +1,7 @@
 #ifndef PASCAL_OPT_H
 #define PASCAL_OPT_H
 
-#include "Pascal/ast.h"
+#include "ast/ast.h"
 
 // Perform simple AST optimizations such as constant folding and
 // dead-branch elimination. Returns the potentially replaced root node.
