+++ src/ast/ast.c
@@ -1,8 +1,8 @@
 #include "core/types.h"
 #include "core/utils.h"
 #include "core/list.h"
-#include "Pascal/ast.h"
-#include "globals.h"
+#include "ast.h"
+#include "Pascal/globals.h"
 #include "symbol/symbol.h"
 #include "Pascal/parser.h"
 #include "backend_ast/builtin.h"
