+++ src/rea/type_stubs.c
@@ -1,0 +1,13 @@
+#include "ast/ast.h"
+#include "core/types.h"
+#include "core/utils.h"
+
+AST* lookupType(const char* name) {
+    (void)name;
+    return NULL;
+}
+
+void insertType(const char* name, AST* typeDef) {
+    (void)name;
+    (void)typeDef;
+}
